+++ config/database.php
@@ -102,8 +102,12 @@ return [
             'charset' => env('DB_CHARSET', 'utf8'),
             'prefix' => '',
             'prefix_indexes' => true,
-            'search_path' => 'public',
+            'search_path' => env('DB_SCHEMA', 'public'),
             'sslmode' => env('DB_SSLMODE', 'prefer'),
+            'options' => extension_loaded('pdo_pgsql')
+                && filter_var(env('DB_PGSQL_EMULATE_PREPARES', false), FILTER_VALIDATE_BOOLEAN)
+                    ? [PDO::ATTR_EMULATE_PREPARES => true]
+                    : [],
         ],
 
         'sqlsrv' => [